requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
redis==5.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4